requirements.txt
load-dotenv==0.1.0
pillow==11.1.0
python-dotenv==1.0.1
schedule==1.2.2
